Create custom plugin for desktop bins building
