+++ jvm/buildSrc/build.gradle.kts
@@ -1,0 +1,16 @@
+gradlePlugin {
+    plugins {
+        create("cargo-desktop") {
+            id = "com.radixdlt.cargo.desktop"
+            implementationClass = "com.radixdlt.cargo.desktop.DesktopCargoPlugin"
+        }
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+plugins {
+    `kotlin-dsl`
+}
